+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Added user result functionality
 * Added possibility to add quiz to favorites
+* Added possibility to buy not free quizzes
 
 ## v0.5.0 - 10.06.2021